Move pagination machinery to avoid circular import
